+++ regions/shapes/polygon.py
@@ -3,10 +3,12 @@
 This module defines polygon regions in both pixel and sky coordinates.
 """
 
+import astropy.units as u
 from astropy.wcs.utils import pixel_to_skycoord, skycoord_to_pixel
 import numpy as np
 
-from ..core.attributes import OneDPix, OneDSky
+from ..core.attributes import (OneDPix, OneDSky, ScalarPix, ScalarLength,
+                               QuantityLength)
 from ..core.bounding_box import BoundingBox
 from ..core.core import PixelRegion, SkyRegion
 from ..core.mask import RegionMask
@@ -15,7 +17,8 @@ from ..core.pixcoord import PixCoord
 from .._geometry import polygonal_overlap_grid
 from .._geometry.pnpoly import points_in_polygon
 
-__all__ = ['PolygonPixelRegion', 'PolygonSkyRegion']
+__all__ = ['PolygonPixelRegion', 'RegularPolygonPixelRegion',
+           'PolygonSkyRegion']
 
 
 class PolygonPixelRegion(PixelRegion):
@@ -191,6 +194,145 @@ class PolygonPixelRegion(PixelRegion):
         """
         vertices = self.vertices.rotate(center, angle)
         return self.copy(vertices=vertices)
+
+
+class RegularPolygonPixelRegion(PolygonPixelRegion):
+    """
+    A regular polygon in pixel coordinates.
+
+    .. note::
+
+        This class will be serialized as a generic polygon
+        region, thus when read back in it will produce a
+        `~regions.PolygonPixelRegion` object instead of a
+        `~regions.RegularPolygonPixelRegion` object.
+
+    Parameters
+    ----------
+    center : `~regions.PixCoord`
+        The position of the center of the polygon.
+    nvertices : `~regions.PixCoord`
+        The number of polygon vertices (or sides).
+    radius : float
+        The distance from the center to any vertex. This is also known
+        as the circumradius.
+    angle : `~astropy.units.Quantity`, optional
+        The rotation angle of the polygon, measured anti-clockwise. If
+        set to zero (the default), the polygon will point "up" following
+        the `matplotlib.patches.RegularPolygon` convention.
+    meta : `~regions.RegionMeta`, optional
+        A dictionary that stores the meta attributes of this region.
+    visual : `~regions.RegionVisual`, optional
+        A dictionary that stores the visual meta attributes of this
+        region.
+
+    Attributes
+    ----------
+    side_length : float
+        The side length.
+    inradius : float
+        The radius of the largest circle contained entirely within the
+        polygon. This value is identical to the length of a line segment
+        from the polygon center to the midpoint of one of its sides
+        (known as as the apothem).
+    perimeter : float
+        The polygon perimeter.
+    interior_angle : float
+        The polygon interior angle, which is the angle at each vertex on
+        the inside of the polygon.
+    exterior_angle : float
+        The polygon exterior angle, which is an angle at each vertex on
+        the ouside of the polygon.
+
+    Examples
+    --------
+    .. plot::
+        :include-source:
+
+        import astropy.units as u
+        import matplotlib.pyplot as plt
+        from regions import PixCoord, RegularPolygonPixelRegion
+
+        fig, ax = plt.subplots(1, 1)
+        center = PixCoord(x=50, y=50)
+        reg1 = RegularPolygonPixelRegion(center, 6, 15)
+        reg1.plot(edgecolor='red', lw=2)
+
+        center = PixCoord(x=25, y=25)
+        reg2 = RegularPolygonPixelRegion(center, 3, 15)
+        reg2.plot(edgecolor='green', lw=2)
+
+        center = PixCoord(x=25, y=75)
+        reg3 = RegularPolygonPixelRegion(center, 3, 15, angle=25*u.deg)
+        reg3.plot(edgecolor='orange', lw=2)
+
+        center = PixCoord(x=75, y=75)
+        reg4 = RegularPolygonPixelRegion(center, 8, 15)
+        reg4.plot(edgecolor='blue', lw=2)
+
+        center = PixCoord(x=75, y=25)
+        reg5 = RegularPolygonPixelRegion(center, 5, 15)
+        reg5.plot(edgecolor='magenta', lw=2)
+
+        ax.set_xlim(0, 100)
+        ax.set_ylim(0, 100)
+        ax.set_aspect('equal')
+    """
+
+    _params = ('center', 'nvertices', 'radius', 'angle')
+    center = ScalarPix('center')
+    nvertices = ScalarLength('nvertices')
+    radius = ScalarLength('radius')
+    angle = QuantityLength('angle')
+
+    def __init__(self, center, nvertices, radius, angle=0. * u.deg,
+                 meta=None, visual=None):
+
+        if nvertices < 3:
+            raise ValueError('nvertices must be >= 3')
+        self.center = center
+        self.nvertices = nvertices
+        self.radius = radius
+        self.angle = angle
+
+        super().__init__(self._calc_vertices(), meta=meta, visual=visual)
+
+        self.side_length = 2. * self.radius * np.sin(np.pi / self.nvertices)
+        self.inradius = self.radius * np.cos(np.pi / self.nvertices)
+        self.perimeter = self.side_length * self.nvertices
+        self.interior_angle = ((self.nvertices - 2) / self.nvertices
+                               * 180) << u.deg
+        self.exterior_angle = 360. / self.nvertices << u.deg
+
+    def _calc_vertices(self):
+        # uses the matplotlib convention that the polygon always points "up"
+        theta = ((2. * np.pi / self.nvertices * np.arange(self.nvertices)
+                 + (np.pi / 2)) << u.radian) + self.angle
+        xvert = self.radius * np.cos(theta)
+        yvert = self.radius * np.sin(theta)
+        return self.center + PixCoord(xvert, yvert)
+
+    def rotate(self, center, angle):
+        """
+        Rotate the region.
+
+        Postive ``angle`` corresponds to counter-clockwise rotation.
+
+        Parameters
+        ----------
+        center : `~regions.PixCoord`
+            The rotation center point.
+        angle : `~astropy.coordinates.Angle`
+            The rotation angle.
+
+        Returns
+        -------
+        region : `PolygonPixelRegion`
+            The rotated region (which is an independent copy).
+        """
+        center = self.center.rotate(center, angle)
+        angle = self.angle + angle
+        return self.copy(center=center, angle=angle)
 
 
 class PolygonSkyRegion(SkyRegion):